requirements.txt
python-dotenv
selenium
./undetected-chromedriver-3.5.0.tar.gz